refactor: init folders and renaming structure
