+++ bentoml/_internal/service/service.py
@@ -210,6 +210,7 @@ class Service:
     def on_asgi_app_shutdown(self) -> None:
         pass
 
+    @property
     def grpc_server(self) -> GRPCServer:
         from ..server.grpc_server import GRPCServer
